requirements.txt
openai>=1.40.0
python-dotenv
fastapi
uvicorn
